pyproject.toml
[project]
name = "3b"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.6",
    "guardrails-ai==0.6.0",
    "guardrails-api-client==0.4.0a1",
    "langchain-core>=0.3.22",
    "nltk>=3.9.1",
    "openai>=1.57.0",
    "python-dotenv>=1.0.1",
]
